Configurations files for new merged URDF
model_merged.urdf is a new URDF file for iCubParis02. A hybrid between iCubGenova05's URDF and iCubParis02's. For instance, the upper body properties have been copied from Genova05's and also the new way how sensors are specified. The legs properties remain as originally were, wince Paris02's legs are an older version w.r.t. to Genova05's. 

The file yarpWholeBodyInterface_mergedURDF.ini is the configuration file that we're using when tuning the torque control of Paris02. In particular, it uses model_merged.urdf, the WORLD_REFERENCE_FRAME is specified as l_sole, readSpeedAccFromControlBoard is also active and port prefixes of the controlboards are prepended with jtc/ . In addition, in this config file you will find the additional list ROBOT_SINGLE_JOINT which we are using with the WB-Toolbox to tune the torque control taking advantage of the new model and gravity compensation.