Disable Wall on MSVC
On Visual Studio, `/Wall` is a level of warning that report warning that are not actually warnings 
(see https://msdn.microsoft.com/en-us/library/thxezb7y.aspx). 
To reduce the log in AppVeyor build of the codyco-superbuild , I think we can disable this warnings (even considering that this library should be deprecated sooner or later).